fix(entities): 修正实体模型字段类型和结构体属性
- cti_phone_location实体中添加Eq trait以保证类型完整性
- 将cti_phone_location中字段q的类型由Float改为Decimal以增强精度
- cti_stream_entity实体中为id字段添加声明确保结构体定义完整

+++ docs/sea-orm-entities/cti/cti_phone_location.rs
@@ -3,7 +3,7 @@
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
 #[sea_orm(schema_name = "cti", table_name = "cti_phone_location")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
@@ -16,8 +16,8 @@ pub struct Model {
     pub received: Option<String>,
     pub path: Option<String>,
     pub expires: Option<DateTime>,
-    #[sea_orm(column_type = "Float", nullable)]
-    pub q: Option<f32>,
+    #[sea_orm(column_type = "Decimal(Some((10, 3)))", nullable)]
+    pub q: Option<Decimal>,
     pub callid: Option<String>,
     pub cseq: Option<i32>,
     pub last_modified: Option<DateTime>,
